requirements.txt
Django~=5.0.2
djangorestframework~=3.14.0
cryptography~=42.0.2
pytz~=2024.1
asgiref~=3.7.2
sqlparse~=0.4.4
requests~=2.31.0
django-allauth~=0.61.1
PyJWT~=2.8.0
certifi~=2024.2.2
urllib3~=2.2.0
oauthlib~=3.2.2
idna~=3.6
pycparser~=2.21
django-rest-auth~=0.9.5
six~=1.16.0